sets `HeadlineAlt` to `headline` if it exists
Prior to this commit, `Headline` is set to
`alert["properties"]["parameters"]["NWSheadline"]. However, the
human-readable string on weather.gov uses the value of
`properties.headline`. It would be nice to have both available in
`nws_alerts` sensor.

This commit maintains the existing behavior of setting `Headline` to
`NWSheadline` if it exists, or `event` if it doesn't. It introduces
a new `HeadlineAlt` key in the dict, set to `properties.headline`,
if it exists in the response. If it doesn't exist, `HeadlineAlt` is
set to the value of `Headline`.

Examples:

NWSheadline:
    FLOOD WARNING NOW IN EFFECT UNTIL LATE TOMORROW EVENING

event:
    Flood Warning

headline:
    Flood Warning issued February 26 at 3:07PM PST until
    February 27 at 11:12PM PST by NWS Portland OR